Revert "OSIS-4652 restructure learning unit form tests"
This reverts commit ccab2b1d7039f944ac76e7ba1158083dac9491c6.
